refatoração da impressão para tirar arquivos desnecessários

+++ SistemaConferencia/Etiquetas/compileJasper.php
@@ -1,5 +1,4 @@
 <?php
-
 require 'C:\xampp\htdocs\vendor\autoload.php';
 
 use PHPJasper\PHPJasper;
@@ -9,6 +8,8 @@ $funcao = $_POST['funcao'];
 if ($funcao === 'compileJasper') {
     compileJasper();
 } else if ($funcao === 'fechaJanelaWcpp') {
+    // Fechar a janela do WebClientPrint é tratado de forma diferente no Windows e no Linux.
+    // Para o Linux, esperamos x segundos para não matar o processo antes que ele envie os dados para a impressora.
     usleep(2500000);
     fechaJanelaWcpp();
 }
@@ -16,6 +17,7 @@ if ($funcao === 'compileJasper') {
 
 function compileJasper()
 {
+    // A biblioteca externa PHPJasper transforma o arquivo .jrxml gerado pelo iReport para um pdf
     $param = $_POST['nunota'];
     $file = $_POST['arquivo'];
 
@@ -25,6 +27,7 @@ function compileJasper()
 
     $jasper = new PHPJasper;
 
+    // Transforma o arquivo .jrxml em um arquivo .jasper
     $jasper->compile(
         $input,
         $output
@@ -34,6 +37,8 @@ function compileJasper()
     $output = 'C:\xampp\htdocs\SistemaConferencia\Etiquetas\nunotas/' . $param;
     $jdbc_dir = 'C:\xampp\htdocs\vendor\geekcom\phpjasper\bin\jasperstarter\jdbc';
 
+    // Opções para a conexão com o banco, para que o relatório seja gerado com dados atualizados.
+    // Importante configurar certo o jdbc_dir, no momento é usado o sqljdbc4 para o SQL Server   
     $options = [
         'format' => ['pdf'],
         'locale' => 'en',
@@ -53,6 +58,7 @@ function compileJasper()
         ]
     ];
 
+    // Transforma o arquivo .jasper em um arquivo .pdf, que será usado para a impressão
     $jasper->process(
         $input,
         $output,
@@ -62,18 +68,24 @@ function compileJasper()
 
 function fechaJanelaWcpp()
 {
-    $host = $_SERVER['REMOTE_ADDR'];
-    $port = 22;
-    $username = 'administrador';
-    $password = 'G3r#2oi5';
+    $user_agent = $_SERVER['HTTP_USER_AGENT'];
+    // Verifica se é um cliente Linux fazendo a requisição
+    if (strpos($user_agent, 'Linux') !== false) {
+        $host = $_SERVER['REMOTE_ADDR'];
+        $port = 22;
+        $username = 'administrador';
+        $password = 'G3r#2oi5';
+        // Conecta por SSH com o cliente Linux
+        $ssh = new SSH2($host, $port);
+        if (!$ssh->login($username, $password)) {
+            exit('Falha na autenticação');
+        }
+        // Executa um comando para matar o processo da janela do WebClientPrint
+        $comando = 'echo G3r#2oi5 | sudo -S killall -e wcpp6';
+        $output = $ssh->exec($comando);
 
-    $ssh = new SSH2($host, $port);
-    if (!$ssh->login($username, $password)) {
-        exit('Falha na autenticação');
+        echo $output;
+        // Para configurar a conexão SSH com o cliente Linux, rodar os seguintes comandos no terminal: "apt install openssh-server" e "systemctl start ssh"
+        // Comando opcional para verificar se o OpenSSH Server está funcionando: "systemctl status ssh"
     }
-
-    $comando = 'echo G3r#2oi5 | sudo -S killall -e wcpp6';
-    $output = $ssh->exec($comando);
-
-    echo $output;
 }
